backend/requirements.txt
django
django_json_widget
google-genai
djangorestframework
djangorestframework-simplejwt
django-environ
django-cors-headers
django-filter
gunicorn
PyYAML
